Updating tests for refactored curriculum learning.
